+++ src/components/responsive/hide_from.test.js
@@ -2,15 +2,17 @@ import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test';
 
-import { EuiHideFrom } from './hide_from';
+import { RESPONSIVE_SIZES, EuiHideFor } from './hide_from';
 
-describe('EuiHideFrom', () => {
-  test('is rendered', () => {
-    const component = render(
-      <EuiHideFrom sizes={['xs', 's', 'm', 'l']} {...requiredProps} />
-    );
+describe('EuiHideFor', () => {
+  RESPONSIVE_SIZES.forEach(size => {
+    test(`${size} is rendered`, () => {
+      const component = render(
+        <EuiHideFor sizes={[size]} {...requiredProps} />
+      );
 
-    expect(component)
-      .toMatchSnapshot();
+      expect(component)
+        .toMatchSnapshot();
+    });
   });
 });