+++ src/components/responsive/show_for.test.js
@@ -2,15 +2,17 @@ import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test';
 
-import { EuiShowFor } from './show_for';
+import { RESPONSIVE_SIZES, EuiShowFor } from './show_for';
 
 describe('EuiShowFor', () => {
-  test('is rendered', () => {
-    const component = render(
-      <EuiShowFor sizes={['xs', 's', 'm', 'l']} {...requiredProps} />
-    );
+  RESPONSIVE_SIZES.forEach(size => {
+    test(`${size} is rendered`, () => {
+      const component = render(
+        <EuiShowFor sizes={[size]} {...requiredProps} />
+      );
 
-    expect(component)
-      .toMatchSnapshot();
+      expect(component)
+        .toMatchSnapshot();
+    });
   });
 });